+++ sdk/src/main/java/com/twincoders/twinpush/sdk/services/NotificationIntentService.java
@@ -1,28 +1,17 @@
 package com.twincoders.twinpush.sdk.services;
 
-import android.app.Notification;
 import android.app.PendingIntent;
 import android.content.Context;
-import android.content.Intent;
 
 import androidx.annotation.NonNull;
-import androidx.core.app.NotificationCompat;
 import androidx.core.app.NotificationManagerCompat;
 
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
-import com.twincoders.twinpush.sdk.R;
 import com.twincoders.twinpush.sdk.TwinPushSDK;
 import com.twincoders.twinpush.sdk.logging.Ln;
-import com.twincoders.twinpush.sdk.logging.Strings;
 import com.twincoders.twinpush.sdk.notifications.PushNotification;
 
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.Map;
 
 public class NotificationIntentService extends FirebaseMessagingService implements PushReceiverService {
@@ -54,8 +43,12 @@ public class NotificationIntentService extends FirebaseMessagingService implemen
 			// Send push notification acknowledgement if enabled
 			if (twinpush.isPushAckEnabled() && NotificationManagerCompat.from(getBaseContext()).areNotificationsEnabled())
 				twinpush.onNotificationReceived(notification);
-			// Display Notification
-			displayNotification(getBaseContext(), notification);
+			// Display Notification if not silent
+			if (notification.isSilent()) {
+				onSilentPushReceived(getBaseContext(), notification);
+			} else {
+				displayNotification(getBaseContext(), notification);
+			}
 		}
     }
 
@@ -69,7 +62,7 @@ public class NotificationIntentService extends FirebaseMessagingService implemen
         // Get updated InstanceID token.
         Ln.d("FCM Token created: " + s);
 
-        // Refresh register if needed
+        // Refresh registration if needed
         if (twinPush.isDeviceRegistered()) {
             twinPush.register();
         }
@@ -90,5 +83,10 @@ public class NotificationIntentService extends FirebaseMessagingService implemen
 	@Override
 	public PushNotification getNotification(Map<String, String> data) {
 		return defaultService.getNotification(data);
+	}
+
+	@Override
+	public void onSilentPushReceived(Context context, PushNotification notification) {
+    	defaultService.onSilentPushReceived(context, notification);
 	}
 }